tests/smoke/examples/synthetic/app95_marketing_mix_model/requirements.txt
streamlit>=1.28
pandas
numpy
gspread>=5.0
google-cloud-bigquery>=3.0
google-auth
